+++ RestAPI/functions/controllers/methods/getMethods.js
@@ -1,4 +1,5 @@
 const { dataConnect } = require("../../config/firebase.js");
+const { checkMethodExists } = require("./methodUtils.js");
 
 const getAllMethods = async (req, res) => {
   try {
@@ -101,45 +102,7 @@ const getOneMethod = async (req, res) => {
   }
 };
 
-const checkMethodExists = async (methodId) => { 
-    try {
-        if (!methodId) {
-          throw new Error("methodId is required");
-        }
-
-        const variables = {methodId};
-        const GET_METHOD_QUERY = `
-            query GetMethodById($methodId: String!) @auth(level: USER) {
-                method(key: {id: $methodId}) {
-                    id
-                    name
-                    description
-                    price
-                    createdAt
-                    updatedAt
-                }
-            }
-        `;
-        
-        console.log("Executing GraphQL query:", GET_METHOD_QUERY, "with variables:", variables);
-        const response = await dataConnect.executeGraphql(GET_METHOD_QUERY, {
-          variables: variables,
-        });
-        
-        respondData = response.data.method;
-        console.log("Response data:", respondData);
-
-        if (!respondData) {
-            return false;
-        } else return true;
-    } catch (error) {
-        console.error("Error checking method existence:", error);
-        throw new Error("Failed to check method existence");
-    }
-}
-
 module.exports = {
   getAllMethods,
   getOneMethod,
-  checkMethodExists,
 };
